+++ Cmdr/BuiltInTypes/PlayerId.lua
@@ -22,6 +22,7 @@ end
 
 local playerIdType = {
 	DisplayName = "Full Player Name";
+	Prefixes = "# integer";
 
 	Transform = function (text)
 		local findPlayer = Util.MakeFuzzyFinder(Players:GetPlayers())
@@ -44,5 +45,7 @@ local playerIdType = {
 
 return function (cmdr)
 	cmdr:RegisterType("playerId", playerIdType)
-	cmdr:RegisterType("playerIds", Util.MakeListableType(playerIdType))
+	cmdr:RegisterType("playerIds", Util.MakeListableType(playerIdType, {
+		Prefixes = "# integers"
+	}))
 end